+++ opentelemetry-sdk/src/metrics/manual_reader.rs
@@ -12,10 +12,7 @@ use super::{
     data::{ResourceMetrics, Temporality},
     instrument::InstrumentKind,
     pipeline::Pipeline,
-    reader::{
-        AggregationSelector, DefaultAggregationSelector, DefaultTemporalitySelector,
-        MetricProducer, MetricReader, SdkProducer, TemporalitySelector,
-    },
+    reader::{DefaultTemporalitySelector, MetricReader, SdkProducer, TemporalitySelector},
 };
 
 /// A simple [MetricReader] that allows an application to read metrics on demand.
@@ -34,7 +31,6 @@ use super::{
 pub struct ManualReader {
     inner: Box<Mutex<ManualReaderInner>>,
     temporality_selector: Box<dyn TemporalitySelector>,
-    aggregation_selector: Box<dyn AggregationSelector>,
 }
 
 impl Default for ManualReader {
@@ -53,7 +49,6 @@ impl fmt::Debug for ManualReader {
 struct ManualReaderInner {
     sdk_producer: Option<Weak<dyn SdkProducer>>,
     is_shutdown: bool,
-    external_producers: Vec<Box<dyn MetricProducer>>,
 }
 
 impl ManualReader {
@@ -63,19 +58,13 @@ impl ManualReader {
     }
 
     /// A [MetricReader] which is directly called to collect metrics.
-    pub(crate) fn new(
-        temporality_selector: Box<dyn TemporalitySelector>,
-        aggregation_selector: Box<dyn AggregationSelector>,
-        producers: Vec<Box<dyn MetricProducer>>,
-    ) -> Self {
+    pub(crate) fn new(temporality_selector: Box<dyn TemporalitySelector>) -> Self {
         ManualReader {
             inner: Box::new(Mutex::new(ManualReaderInner {
                 sdk_producer: None,
                 is_shutdown: false,
-                external_producers: producers,
             })),
             temporality_selector,
-            aggregation_selector,
         }
     }
 }
@@ -83,12 +72,6 @@ impl ManualReader {
 impl TemporalitySelector for ManualReader {
     fn temporality(&self, kind: InstrumentKind) -> Temporality {
         self.temporality_selector.temporality(kind)
-    }
-}
-
-impl AggregationSelector for ManualReader {
-    fn aggregation(&self, kind: InstrumentKind) -> super::aggregation::Aggregation {
-        self.aggregation_selector.aggregation(kind)
     }
 }
 
@@ -108,7 +91,7 @@ impl MetricReader for ManualReader {
         });
     }
 
-    /// Gathers all metrics from the SDK and other [MetricProducer]s, calling any
+    /// Gathers all metrics from the SDK, calling any
     /// callbacks necessary and returning the results.
     ///
     /// Returns an error if called after shutdown.
@@ -123,19 +106,7 @@ impl MetricReader for ManualReader {
             }
         };
 
-        let mut errs = vec![];
-        for producer in &inner.external_producers {
-            match producer.produce() {
-                Ok(metrics) => rm.scope_metrics.push(metrics),
-                Err(err) => errs.push(err),
-            }
-        }
-
-        if errs.is_empty() {
-            Ok(())
-        } else {
-            Err(MetricsError::Other(format!("{:?}", errs)))
-        }
+        Ok(())
     }
 
     /// ForceFlush is a no-op, it always returns nil.
@@ -150,7 +121,6 @@ impl MetricReader for ManualReader {
         // Any future call to collect will now return an error.
         inner.sdk_producer = None;
         inner.is_shutdown = true;
-        inner.external_producers = Vec::new();
 
         Ok(())
     }
@@ -159,8 +129,6 @@ impl MetricReader for ManualReader {
 /// Configuration for a [ManualReader]
 pub struct ManualReaderBuilder {
     temporality_selector: Box<dyn TemporalitySelector>,
-    aggregation_selector: Box<dyn AggregationSelector>,
-    producers: Vec<Box<dyn MetricProducer>>,
 }
 
 impl fmt::Debug for ManualReaderBuilder {
@@ -173,8 +141,6 @@ impl Default for ManualReaderBuilder {
     fn default() -> Self {
         ManualReaderBuilder {
             temporality_selector: Box::new(DefaultTemporalitySelector { _private: () }),
-            aggregation_selector: Box::new(DefaultAggregationSelector { _private: () }),
-            producers: vec![],
         }
     }
 }
@@ -196,35 +162,8 @@ impl ManualReaderBuilder {
         self
     }
 
-    /// Sets the [AggregationSelector] a reader will use to determine the
-    /// aggregation to use for an instrument based on its kind.
-    ///
-    /// If this option is not used, the reader will use the default aggregation
-    /// selector or the aggregation explicitly passed for a view matching an
-    /// instrument.
-    pub fn with_aggregation_selector(
-        mut self,
-        aggregation_selector: impl AggregationSelector + 'static,
-    ) -> Self {
-        self.aggregation_selector = Box::new(aggregation_selector);
-        self
-    }
-
-    /// Registers a an external [MetricProducer] with this reader.
-    ///
-    /// The producer is used as a source of aggregated metric data which is
-    /// incorporated into metrics collected from the SDK.
-    pub fn with_producer(mut self, producer: impl MetricProducer + 'static) -> Self {
-        self.producers.push(Box::new(producer));
-        self
-    }
-
     /// Create a new [ManualReader] from this configuration.
     pub fn build(self) -> ManualReader {
-        ManualReader::new(
-            self.temporality_selector,
-            self.aggregation_selector,
-            self.producers,
-        )
+        ManualReader::new(self.temporality_selector)
     }
 }